Make the app grid horizontal
As per design decision, make the app grid horizontal. This
is implemented through the 'teardown & rebuild' approach
where all related actors are initially removed from their
parents, and then readded to the new location.

In the case of PagesIndicator, it is necessary to create
a new instance, because BaseAppView uses different classes
depending on the orientation.

The scrollbar adjustment handling required some handholding
and (reasonably) assumes that the vertical adjustment won't
change when the scroll view's vscroll policy is NONE.

https://phabricator.endlessm.com/T30544