Import minified IconGridLayout
This used to be the main entry point for Endless OS
app grid management. Import a minified version of it,
leaving out all modification-related code (such as
adding, appending, modifying, or removing icons).

Another difference is that it isn't a singleton anymore;
instead, it's meant to be created in a throwaway fashing,
and the GSetting where it'll read from is passed as a
construct-only property.