layout: Add overview clones to desktop
Implement a new OverviewClone class that clones the
search bar, and creates a fake AppDisplay.

This follows the old approach of creating everything,
and hiding the entire actor tree from picking.

When the eos-panel extension is ready, we'll need to
update this code to add the panel *after* the app grid.

https://phabricator.endlessm.com/T30543