layout: Force disconnect from overview when hidden
Disabling the extension destroys the app grid clones, which
also destroys the AppDisplay instances. However, AppDisplay
doesn't disconnect itself from Main.overview, which is pretty
sad for us.

I'd like to stress out that this is a massive, life shattering
hack, and that I have a special place in all concepts of hell
for writing this code.