layout: Use a real entry in grid clones
This is a complicated problem to explain, so please bear with
me. So, ClutterClones. They redirect the painting to the source
actor, and sometimes the allocation.

For painting to happen, actors need to be mapped. Mapping an
actor for a clone is tricky, since it has to temporarily map
all unmapped parents too. Remapping and unmapping can, sometimes,
cause the projection matrix to be invalidated, and naturally it
needs to be recomputed. Recomputing the actor's projection matrix
is computationally expensive.

Using a ClutterClone for the search entry was a terrible idea,
I should have known better :/  It caused continuous invalidations
of pretty much all actors between and including the app grid,
and the overview actor, all the time.

https://phabricator.endlessm.com/T30788