appDisplay: Don't implicitly return
When using arrow functions, omitting the curly braces
makes JavaScript implicitly return the value. As it
turns out, assigning a variable also means the new
value is returned.

This is problematic here because Overview.Overview is
*not* a GObject, and uses Signal.addSignalMethods() to
be able to handle signals, and GJS treats returning
'true' as 'stop signal emission entirely' [1]. This
ends up preventing the callback of layout.js from
being called, and the animation from the app grid to
the windows group is broken.

Add curly braces to the arrow callbacks, which prevents
JavaScript from implicitly returning.

[1] https://gitlab.gnome.org/GNOME/gjs/-/blob/master/modules/core/_signals.js#L137