viewSelector: disable bgMenu click on search bar
Clicking in the secondary-icon-button on the search bar is triggering
the long-press event for the bgMenuClickAction.

This patch release the background menu click action when this button is
clicked, so the menu is not shown.

This is a workaround to fix this issue, but maybe the best way to solve
this problem is to detect why the long-press event is emitted when the
click is a one click.

https://phabricator.endlessm.com/T30860