layout: Make grid dirty when clone is mapped
This is, once again, another case of FBO contents not
being invalidated properly. In this case, what happens
is that the icon grid doesn't relayout when unmapped,
which skips marking it as dirty, which doesn't trigger
an FBO invalidation and redraw on both the offscreen
cache, and (most importantly) the desaturation effect.

This can be verified by disabling the desaturation
effect and noticing that invalidation happens on one
of the offscreen caches, and thus triggering a redraw
of the grid.

This is ultimately a problem in the paint machinery of
Clutter, but we can easily work around it by forcing the
actor to be marked as dirty after all parents are mapped.

https://phabricator.endlessm.com/T30964