layout: Also bind 'upper' on adjustments
The 'upper' property of the clones' adjustments are only
updated after the first time it's mapped. However, at this
point, the real app grid itself may be on another page. The
clones can't reflect that because their adjustments' 'upper'
property is not yet updated with the number of pages.

Bind the 'upper' property of the real app grid's adjustments
with the clones'. This is fine because the number of pages
is decided based on the number of icons, not in the available
space.

https://phabricator.endlessm.com/T30964