appDisplay: Add navigation arrows
As per design direction, add navigation arrows. The arrows
are actually using swipe arrows to signify that the grid
can also be swipped.

The back arrow is shown starting the middle of the first
page, and vice versa for the next arrow and last page. This
turned out to be a less disruptive way to manage these
arrows.

https://phabricator.endlessm.com/T32384