Show Dash in session mode
This is a moderately tricky implementation of a dock, but I
think it's solid. It reuses the Dash when in session mode, and
adds a new container that mimics what ControlsManagerLayout
does. While the Dash is in the session container, inject a fake
Dash that does nothing.

There are other aspects of a dock-like Dash that need to be
figured out, which the following commits will do.