dash: Hide when windows overlap
This effectively introduces the so called IntelliHide
behaviour in the Dash code. Whenever any visible window
"touches" the Dash, it shies away at the bottom of the
screen. Because the parent container is clipped to the
allocation, the Dash never leaks to monitors beneath it.

This is inspired in the Dash-to-Dock code, though it
quite significantly diverges from it.