dash: Workaround actor reparenting issues
This is a small fallout from the approach we've taken here,
sometimes we reparent the Dash before the translation-y
animation ends, in which case we get a harmless warning about
timelines, and the Dash doesn't reach the target value of the
animation.

Fix that by forcefully setting the target translation-y value
when the animation is stopped.