dash: Update barrier when monitor layout changes
We need to account that the monitor layout state when the extension
is enable may change, otherwise the cursor barrier can get stuck in
weird places.

Do that by accounting for Main.layoutManager's 'monitors-changed'
signal, and global.display's 'workareas-changed' signals. The
combination of these two signals is apparently enough to cover all
cases where the cursor barrier needs to be recalculated.

https://phabricator.endlessm.com/T33055