workspaceMonitor: Throttle window updates
WorkspaceMonitor reacts to actors being added to, and
removed from, the window group. However - and this is
something me and Andre discussed years ago - this
approach can backfire when windows are added and removed
in quick succession.

So far we didn't have to deal with this case, but
recently, after fixing other bugs elsewhere in the
extension, Hack started to hit it.

Throttling updates is not a perfect solution by any
means, but it's certainly more robust than reacting
to every single actor addition and removal unconditionally.

https://phabricator.endlessm.com/T33294