dash: Allow clicking through empty areas
This allows clicking through the empty areas surrounding the Dash,
and potentially interact with apps beneath it, and still show the
Dash.

Change the hover check to observe Main.overview.dash._dashContainer,
which is what actually contains the icons. That limits hovering to
the inner area within Dash that contains icons, but not the entire
Dash covering the screen width.

https://phabricator.endlessm.com/T33219