Reuse translations from Shell where possible
These three strings are all present and translated into many languages
in GNOME Shell itself. By accessing them from the 'gnome-shell' gettext
domain, we can reuse those translations. By using this wrapper function
that does not match the patterns passed to xgettext by meson, these
strings will not be extracted for translation, saving on unnecessary
work.

With this change, the only strings remaining in this module are for the
internet search provider.

https://phabricator.endlessm.com/T33603