Drop the icon grid clone
After the big GNOME 40 overview redesign, a lot of the assumptions
that motivated the introduction of the icon grid clone are not
appliable anymore. In fact, recent user studies shows that users
either do not notice the app grid clone laying behind windows, or
that it actively confuses them.

Let's see how we go without it.

https://phabricator.endlessm.com/T33947