Embolden font and sharpen shadow on app grid labels
Prior to this commit, the font was at regular weight, with a 1px vertical offset,
3px radius and 60% opacity black shadow.

Embolden the font & make the shadow opaque black, with a 2px vertical
offset and 2px radius. This matches the settings we had on EOS 4 and
makes the app grid legible against the wallpaper now that we have
stopped darkening it.

https://phabricator.endlessm.com/T34129