panel: Clean up transitions between UI modes
With this change, clicking the Activities button while the Activities
overview is active will consistently exit that mode in one of two
fashions. If any windows are open, it will show those windows. If not,
the shell will switch to the Applications grid.

In addition, the "Show Apps" button in the dash now behaves consistently
with the "Applications" button in the panel. Clicking it when the
Applications grid is visible will show either the current open windows,
or the overview.

https://phabricator.endlessm.com/T35108