search: Fix the overview staying open on activate
In 2b1f80d6d184994e35798357a7153902f974cba0, we added code to force the
overview to hide when a search result is activated. However, this was
incorrectly overriding a (nonexistent) `hide()` function in
`Search.SearchResult`. As a result, we have bene using the default
behaviour the `activate()` function. It calls `Main.overview.toggle()`,
which hides the overview as long as at least one window is visible,
giving the illusion that it was working as expected.

Instead, override `activate()` and force the overview to hide even if
there are no windows visible at the moment. In addition, provide the
same treatment for the signal handler which runs when a search provider
is clicked.

https://phabricator.endlessm.com/T35132