Name all overridden methods
Previously, when a method had been overridden with Utils.override(), it
would show up in stack traces without any name. For example, in the
following stack trace:

    JS ERROR: Error: Invalid overview shown transition from SHOWING to SHOWING
    _changeShownState@resource:///org/gnome/shell/ui/overview.js:295:25
    runStartupAnimation@resource:///org/gnome/shell/ui/overview.js:686:14
    enable/<@/usr/share/gnome-shell/extensions/eos-desktop@endlessm.com/ui/overview.js:96:24
    _startupAnimationSession@resource:///org/gnome/shell/ui/layout.js:759:33
    _startupAnimation@resource:///org/gnome/shell/ui/layout.js:743:24
    _prepareStartupAnimation@resource:///org/gnome/shell/ui/layout.js:731:24
    async*_loadBackground/signalId</id<@resource:///org/gnome/shell/ui/layout.js:661:26

the
enable/<@/usr/share/gnome-shell/extensions/eos-desktop@endlessm.com/ui/overview.js:96:24
frame is an override of Overview.runStartupAnimation(), with the `enable/<`
notation meaning “an anonymous function defined within enable”.

The names of functions in JavaScript is surprisingly complicated.
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/name
has the gory details on the many ways that functions can be named.  The
relevant point is that our overrides are anonymous functions which are
not assigned directly to a variable or object property, but passed
directly to another function. As a result no name can be inferred and
they show up in stack traces with unhelpful names as above.

Perhaps surprisingly, you can name a function expression:

    let f = function g() { };

In this contrived example, even though you can't invoke the function as
`g()`, `f.name` is `"g"`.

Use this syntax to give names to our overridden methods. To avoid
needing to write out the method name twice, adapt Utils.override() to
figure out the method name to override from the `.name` property of the
given function.

With this change, the stack trace above looks like this:

    JS ERROR: Error: Invalid overview shown transition from SHOWING to SHOWING
    _changeShownState@resource:///org/gnome/shell/ui/overview.js:295:25
    runStartupAnimation@resource:///org/gnome/shell/ui/overview.js:686:14
    runStartupAnimation@/usr/share/gnome-shell/extensions/eos-desktop@endlessm.com/ui/overview.js:96:24
    _startupAnimationSession@resource:///org/gnome/shell/ui/layout.js:759:33
    _startupAnimation@resource:///org/gnome/shell/ui/layout.js:743:24
    _prepareStartupAnimation@resource:///org/gnome/shell/ui/layout.js:731:24
    async*_loadBackground/signalId</id<@resource:///org/gnome/shell/ui/layout.js:661:26

with `enable/<` replaced with `runStartupAnimation`.

https://phabricator.endlessm.com/T34453