Simplify calling original methods
When chaining up to the original method in our overridden
implementations, we have to ensure that `this` is set correctly in the
original method. Previously, the pattern used was:

    f.bind(x)(a, b, c);

MDN says:

> The bind() method of Function instances creates a new function that,
> when called, calls this function with its this keyword set to the
> provided value, and a given sequence of arguments preceding any
> provided when the new function is called.

Function also has a `call()` method:

> The call() method of Function instances calls this function with a
> given this value and arguments provided individually.

Rather than calling `bind()`, calling the result immediately and then
discarding the bound function, use `call()`.

https://phabricator.endlessm.com/T34453