Make runStartupAnimation() functions fully async
Previously, in gnome-shell, the functions that handle the startup
animation used a mixture of callbacks and promises to report success.
This made it harder to correctly propagate and handle any exception
raised during the animation.

In gnome-shell, I adjusted these functions to consistently use promises
(either explicitly or via async/await), and so we need to adjust our
overrides here.

https://phabricator.endlessm.com/T34453