Don't try to show app grid during startup animation
We override Overview.hide() in such a way that in some cases it actually
ends up showing the overview. Specifically, if there are no visible
windows, the overridden implementation causes the show-apps button to be
pressed; and a ::notify::checked handler for that button in our
extension causes the overview to be shown in some cases.

This was implemented in commit 248196ec6917b4152859f5aae619e2fccf8a4cd5
for reasons explained there.

One case where Overview.hide() is called is in response to
MetaMonitorManager::monitors-changed, from Overview._relayout(). This
can happen very early in the session, even before the startup animation
has run. This would cause the overview's state to be changed to SHOWING
before the startup animation runs. Then when the startup animation
attempts to change the state to SHOWING, an exception was raised because
SHOWING → SHOWING is not a legal state transition.

Avoid this by not adjusting the show-apps button in the overridden
Overview.hide() method if the startup animation has not yet run.

https://phabricator.endlessm.com/T34453