Revert "appDisplay: Don't switch pages when hiding overview"
Conceptually this reverts commit c1a95cc5b6dbe2d3b33483877c0305d837064c0b.

In vanilla GNOME, when you move away from the app grid back to the session or
window picker, the app grid returns to the first page (if it was not already
there). In Endless OS, we have historically preserved your position on the app
grid: if you scrolled to the second page, then returned to your session, then
returned to the app grid, it would still be at the second page.

This has been broken since at least Endless OS 6: the app grid always loses
your place when you close it.

It took me a little while to understand the comment here, but what's meant to
happen is:

- The overview emits ::hiding; our handler sets its hidingOverview flag to true

- The overview emits ::hidden. There are (at least) two handlers connected:

    a. A handler connected in Shell's AppDisplay._init (i.e. its constructor),
       which calls this.goToPage(0)

    b. Our handler, connected when the extension is enabled, which sets
       hidingOverview to false

- We override goToPage() to be a no-op if hidingOverview is true.

The intention of the code removed in this patch is that our handler (b) is
connected before Shell's own handler (a), because GObject calls signal handlers
in the order they are connected.

I have to confess that I don't fully understand how this previously worked;
today it does not.

While it would still be nice to have this behaviour (and ideally follow the
behaviour seen on some Android devices where the app grid remembers your
position for a while, then reverts to the first page), it is less important now
that the app grid clone is gone, so you don't see the grid returning to the
first page in the background.

Remove the behaviour.