overviewControls: Restore AppDisplay opacity when disabled
In this extension, we vary the opacity of the app grid to fade it in and out as
you enter and leave the APP_GRID state. As a consequence, when the controls are
in state HIDDEN or WINDOW_PICKER, the AppDisplay actor's opacity is set to 0.

In vanilla Shell, its opacity is only adjusted when transitioning in and out of
search.

As a result, if you disable this extension (while not in the APP_GRID state),
the AppDisplay's opacity remains set to 0 until you trigger a search.

Restore its opacity when disabling the extension.

https://phabricator.endlessm.com/T35677