Revert "appDisplay: Forcefully recreate all icons"
This reverts commit f1287c8770299d62f3d7fbefc9430e6b4661b111.

It is not totally clear from the commit message what issue this was trying to
address. I have some faint memory of it being to do with the real app grid and
its desaturated clone getting out of sync, though I don't remember the details.

Unfortunately it appears that this function leaks all the icons it removes.
Watching the RSS of the gnome-shell process on my system as I disable and
enable the extension (which happens in normal usage when the screen is locked
and unlocked), it grows by 30 MiB with each disable/enable cycle.  I imagine
this leak is proportional to the number of app icons in the grid. I have around
330 app icons on the system in question.

Removing this function cures the leak: RSS does not grow as I repeatedly toggle
the extension. It also makes locking and unlocking the screen very noticably
snappier. And I don't notice any regressions from removing this code.

(I didn't actually investigate the cause of the leak any further but my
assumption is that some C-level reference to the removed items remains, causing
them not to be GCed.)

https://phabricator.endlessm.com/T35677